--- a/code/addons/interactions/src/Panel.tsx
+++ b/code/addons/interactions/src/Panel.tsx
@@ -8,10 +8,10 @@ import {
   STORY_THREW_EXCEPTION,
   UNHANDLED_ERRORS_WHILE_PLAYING,
 } from 'storybook/internal/core-events';
+import { type Call, CallStates, EVENTS, type LogItem } from 'storybook/internal/instrumenter';
 import { useAddonState, useChannel, useParameter } from 'storybook/internal/manager-api';
 
 import { global } from '@storybook/global';
-import { type Call, CallStates, EVENTS, type LogItem } from '@storybook/instrumenter';
 
 import { InteractionsPanel } from './components/InteractionsPanel';
 import { ADDON_ID } from './constants';