--- a/code/addons/interactions/src/components/Interaction.stories.tsx
+++ b/code/addons/interactions/src/components/Interaction.stories.tsx
@@ -1,4 +1,5 @@
-import { CallStates } from '@storybook/instrumenter';
+import { CallStates } from 'storybook/internal/instrumenter';
+
 import type { Meta, StoryObj } from '@storybook/react';
 import { expect, userEvent, within } from '@storybook/test';