--- a/code/addons/interactions/src/components/Interaction.tsx
+++ b/code/addons/interactions/src/components/Interaction.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
 
 import { IconButton, TooltipNote, WithTooltip } from 'storybook/internal/components';
+import { type Call, CallStates, type ControlStates } from 'storybook/internal/instrumenter';
 import { styled, typography } from 'storybook/internal/theming';
 
 import { ListUnorderedIcon } from '@storybook/icons';
-import { type Call, CallStates, type ControlStates } from '@storybook/instrumenter';
 
 import { transparentize } from 'polished';