--- a/code/addons/interactions/src/preview.ts
+++ b/code/addons/interactions/src/preview.ts
@@ -1,6 +1,6 @@
+import { instrument } from 'storybook/internal/instrumenter';
 import type { PlayFunction, StepLabel, StepRunner, StoryContext } from 'storybook/internal/types';
 
-import { instrument } from '@storybook/instrumenter';
 // This makes sure that storybook test loaders are always loaded when addon-interactions is used
 // For 9.0 we want to merge storybook/test and addon-interactions into one addon.
 import '@storybook/test';