--- a/code/addons/test/package.json
+++ b/code/addons/test/package.json
@@ -100,7 +100,6 @@
   "dependencies": {
     "@storybook/global": "^5.0.0",
     "@storybook/icons": "^1.2.12",
-    "@storybook/instrumenter": "workspace:*",
     "@storybook/test": "workspace:*",
     "polished": "^4.2.2",
     "prompts": "^2.4.0",