--- a/code/addons/test/src/components/InteractionsPanel.stories.tsx
+++ b/code/addons/test/src/components/InteractionsPanel.stories.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
+import { CallStates } from 'storybook/internal/instrumenter';
 import { ManagerContext } from 'storybook/internal/manager-api';
 import { styled } from 'storybook/internal/theming';
 
-import { CallStates } from '@storybook/instrumenter';
 import type { Meta, StoryObj } from '@storybook/react';
 import { expect, fn, userEvent, waitFor, within } from '@storybook/test';