--- a/code/addons/test/src/components/MethodCall.tsx
+++ b/code/addons/test/src/components/MethodCall.tsx
@@ -1,9 +1,8 @@
 import type { ReactElement } from 'react';
 import React, { Fragment } from 'react';
 
+import type { Call, CallRef, ElementRef } from 'storybook/internal/instrumenter';
 import { useTheme } from 'storybook/internal/theming';
-
-import type { Call, CallRef, ElementRef } from '@storybook/instrumenter';
 
 import { ObjectInspector } from '@devtools-ds/object-inspector';