--- a/code/addons/test/src/components/Panel.test.ts
+++ b/code/addons/test/src/components/Panel.test.ts
@@ -1,7 +1,7 @@
 // @vitest-environment happy-dom
 import { describe, expect, it } from 'vitest';
 
-import { type Call, CallStates, type LogItem } from '@storybook/instrumenter';
+import { type Call, CallStates, type LogItem } from 'storybook/internal/instrumenter';
 
 import { getInteractions } from './Panel';