--- a/code/addons/test/src/components/Panel.tsx
+++ b/code/addons/test/src/components/Panel.tsx
@@ -8,6 +8,7 @@ import {
   STORY_THREW_EXCEPTION,
   UNHANDLED_ERRORS_WHILE_PLAYING,
 } from 'storybook/internal/core-events';
+import { type Call, CallStates, EVENTS, type LogItem } from 'storybook/internal/instrumenter';
 import {
   useAddonState,
   useChannel,
@@ -17,7 +18,6 @@ import {
 import type { API_StatusValue } from 'storybook/internal/types';
 
 import { global } from '@storybook/global';
-import { type Call, CallStates, EVENTS, type LogItem } from '@storybook/instrumenter';
 
 import { ADDON_ID, STORYBOOK_ADDON_TEST_CHANNEL, TEST_PROVIDER_ID } from '../constants';
 import { InteractionsPanel } from './InteractionsPanel';