--- a/code/addons/test/src/components/StatusBadge.tsx
+++ b/code/addons/test/src/components/StatusBadge.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
 
+import { type Call, CallStates } from 'storybook/internal/instrumenter';
 import { styled, typography } from 'storybook/internal/theming';
-
-import { type Call, CallStates } from '@storybook/instrumenter';
 
 export interface StatusBadgeProps {
   status: Call['status'];