--- a/code/addons/test/src/components/StatusIcon.stories.tsx
+++ b/code/addons/test/src/components/StatusIcon.stories.tsx
@@ -1,4 +1,4 @@
-import { CallStates } from '@storybook/instrumenter';
+import { CallStates } from 'storybook/internal/instrumenter';
 
 import { StatusIcon } from './StatusIcon';