--- a/code/addons/test/src/components/TestDiscrepancyMessage.stories.tsx
+++ b/code/addons/test/src/components/TestDiscrepancyMessage.stories.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
+import { CallStates } from 'storybook/internal/instrumenter';
 import { ManagerContext } from 'storybook/internal/manager-api';
 
-import { CallStates } from '@storybook/instrumenter';
 import type { Meta, StoryObj } from '@storybook/react';
 import { fn } from '@storybook/test';