--- a/code/addons/test/src/components/TestDiscrepancyMessage.tsx
+++ b/code/addons/test/src/components/TestDiscrepancyMessage.tsx
@@ -1,10 +1,9 @@
 import React from 'react';
 
 import { Link } from 'storybook/internal/components';
+import { CallStates } from 'storybook/internal/instrumenter';
 import { useStorybookApi } from 'storybook/internal/manager-api';
 import { styled } from 'storybook/internal/theming';
-
-import { CallStates } from '@storybook/instrumenter';
 
 import { DOCUMENTATION_DISCREPANCY_LINK } from '../constants';