--- a/code/core/scripts/entries.ts
+++ b/code/core/scripts/entries.ts
@@ -48,6 +48,8 @@ export const getEntries = (cwd: string) => {
     define('src/babel/index.ts', ['node'], true),
     define('src/cli/bin/index.ts', ['node'], true),
     define('src/bin/index.ts', ['node'], false),
+
+    define('src/instrumenter/index.ts', ['browser', 'node'], true),
   ];
 };