--- a/code/lib/instrumenter/project.json
+++ b/code/lib/instrumenter/project.json
@@ -1,8 +1,0 @@
-{
-  "name": "instrumenter",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "projectType": "library",
-  "targets": {
-    "build": {}
-  }
-}