--- a/code/lib/instrumenter/tsconfig.json
+++ b/code/lib/instrumenter/tsconfig.json
@@ -1,5 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {},
-  "include": ["src/**/*"]
-}