--- a/code/lib/test/package.json
+++ b/code/lib/test/package.json
@@ -44,7 +44,6 @@
   },
   "dependencies": {
     "@storybook/global": "^5.0.0",
-    "@storybook/instrumenter": "workspace:*",
     "@testing-library/dom": "10.4.0",
     "@testing-library/jest-dom": "6.5.0",
     "@testing-library/user-event": "14.5.2",