Closes #

<!-- If your PR is related to an issue, provide the number(s) above; if it resolves multiple issues, be sure to break them up (e.g. "closes #1000, closes #1001"). -->

<!--

Thank you for contributing to Storybook! Please submit all PRs to the `next` branch unless they are specific to the current release. Storybook maintainers cherry-pick bug and documentation fixes into the `main` branch as part of the release process, so you shouldn't need to worry about this. For additional guidance: https://storybook.js.org/docs/contribute

-->

## What I did

- moved `@storybook/instrumenter` into `storybook/internal/instrumenter`

## Checklist for Contributors

### Testing

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to communicate how to test your changes -->

#### The changes in this PR are covered in the following automated tests:

- [ ] stories
- [ ] unit tests
- [ ] integration tests
- [ ] end-to-end tests

#### Manual testing

_This section is mandatory for all contributions. If you believe no manual test is necessary, please state so explicitly. Thanks!_

<!-- Please include the steps to test your changes here. For example:

1. Run a sandbox for template, e.g. `yarn task --task sandbox --start-from auto --template react-vite/default-ts`
2. Open Storybook in your browser
3. Access X story

-->

### Documentation

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to indicate which documentation has been updated. -->

- [ ] Add or update documentation reflecting your changes
- [ ] If you are deprecating/removing a feature, make sure to update
      [MIGRATION.MD](https://github.com/storybookjs/storybook/blob/next/MIGRATION.md)

## Checklist for Maintainers

- [ ] When this PR is ready for testing, make sure to add `ci:normal`, `ci:merged` or `ci:daily` GH label to it to run a specific set of sandboxes. The particular set of sandboxes can be found in `code/lib/cli-storybook/src/sandbox-templates.ts`
- [ ] Make sure this PR contains **one** of the labels below:
   <details>
     <summary>Available labels</summary>

  - `bug`: Internal changes that fixes incorrect behavior.
  - `maintenance`: User-facing maintenance tasks.
  - `dependencies`: Upgrading (sometimes downgrading) dependencies.
  - `build`: Internal-facing build tooling & test updates. Will not show up in release changelog.
  - `cleanup`: Minor cleanup style change. Will not show up in release changelog.
  - `documentation`: Documentation **only** changes. Will not show up in release changelog.
  - `feature request`: Introducing a new feature.
  - `BREAKING CHANGE`: Changes that break compatibility in some way with current major version.
  - `other`: Changes that don't fit in the above categories.

   </details>

### 🦋 Canary release

<!-- CANARY_RELEASE_SECTION -->
This pull request has been released as version `0.0.0-pr-30740-sha-04411b17`. Try it out in a new sandbox by running `npx storybook@0.0.0-pr-30740-sha-04411b17 sandbox` or in an existing project with `npx storybook@0.0.0-pr-30740-sha-04411b17 upgrade`.
<details>
<summary>More information</summary>

| | |
| --- | --- |
| **Published version** | [`0.0.0-pr-30740-sha-04411b17`](https://npmjs.com/package/storybook/v/0.0.0-pr-30740-sha-04411b17) |
| **Triggered by** | @valentinpalkovic |
| **Repository** | [storybookjs/storybook](https://github.com/storybookjs/storybook) |
| **Branch** | [`valentin/move-instrumenter-into-core`](https://github.com/storybookjs/storybook/tree/valentin/move-instrumenter-into-core) |
| **Commit** | [`04411b17`](https://github.com/storybookjs/storybook/commit/04411b1769845218fda633368402c3ed97a86f3c) |
| **Datetime** | Mon Mar  3 20:03:28 UTC 2025 (`1741032208`) |
| **Workflow run** | [13639398105](https://github.com/storybookjs/storybook/actions/runs/13639398105) |

To request a new release of this pull request, mention the `@storybookjs/core` team.

_core team members can create a new canary release [here](https://github.com/storybookjs/storybook/actions/workflows/canary-release-pr.yml) or locally with `gh workflow run --repo storybookjs/storybook canary-release-pr.yml --field pr=30740`_
</details>
<!-- CANARY_RELEASE_SECTION -->

<!-- BENCHMARK_SECTION -->

| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createSize |  0 B | 0 B | 0 B | - | - |
| generateSize |  79.9 MB | 79.9 MB | 0 B | 0.81 | 0% |
| initSize |  79.9 MB | 79.9 MB | 0 B | 0.81 | 0% |
| diffSize |  97 B | 97 B | 0 B | - | 0% |
| buildSize |  7.45 MB | 7.57 MB | 119 kB | **2.48** | 1.6% |
| buildSbAddonsSize |  1.9 MB | 1.98 MB | 81 kB | **1692.19** | 4.1% |
| buildSbCommonSize |  195 kB | 195 kB | 0 B | - | 0% |
| buildSbManagerSize |  1.88 MB | 1.88 MB | 0 B | -0.33 | 0% |
| buildSbPreviewSize |  0 B | 0 B | 0 B | - | - |
| buildStaticSize |  0 B | 0 B | 0 B | - | - |
| buildPrebuildSize |  3.97 MB | 4.06 MB | 81 kB | **41.72** | 2% |
| buildPreviewSize |  3.47 MB | 3.51 MB | 38 kB | 1.23 | 1.1% |
| testBuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbAddonsSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbCommonSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbManagerSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbPreviewSize |  0 B | 0 B | 0 B | - | - |
| testBuildStaticSize |  0 B | 0 B | 0 B | - | - |
| testBuildPrebuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildPreviewSize |  0 B | 0 B | 0 B | - | - |



| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createTime |  24s | 7.7s | -16s -353ms | -1.19 | -211.4% |
| generateTime |  17.5s | 19.4s | 1.8s | 0.08 | 9.7% |
| initTime |  4.1s | 4.5s | 393ms | 0.04 | 8.7% |
| buildTime |  8.3s | 8.6s | 274ms | -0.39 | 3.2% |
| testBuildTime |  0ms | 0ms | 0ms | - | - |
| devPreviewResponsive |  5.1s | 5s | -65ms | 0.09 | -1.3% |
| devManagerResponsive |  4.9s | 4.9s | -28ms | 0.2 | -0.6% |
| devManagerHeaderVisible |  850ms | 796ms | -54ms | -0.04 | -6.8% |
| devManagerIndexVisible |  903ms | 812ms | -91ms | -0.21 | -11.2% |
| devStoryVisibleUncached |  2.6s | 2.6s | 44ms | 0.7 | 1.6% |
| devStoryVisible |  1.2s | 1.2s | -25ms | 0.35 | -2% |
| devAutodocsVisible |  1.1s | 1.1s | 54ms | 0.76 | 4.6% |
| devMDXVisible |  1.1s | 1.1s | -58ms | 0.25 | -5.2% |
| buildManagerHeaderVisible |  821ms | 813ms | -8ms | -0.09 | -1% |
| buildManagerIndexVisible |  838ms | 830ms | -8ms | -0.13 | -1% |
| buildStoryVisible |  812ms | 803ms | -9ms | -0.04 | -1.1% |
| buildAutodocsVisible |  757ms | 696ms | -61ms | -0.35 | -8.8% |
| buildMDXVisible |  752ms | 797ms | 45ms | 1.01 | 5.6% |

<!-- BENCHMARK_SECTION -->


<!-- greptile_comment -->

## Greptile Summary

Moving the @storybook/instrumenter package into Storybook core, consolidating functionality and updating import paths across the codebase while maintaining existing behavior.

- Relocated instrumenter code from `code/lib/instrumenter` to `code/core/src/instrumenter` with updated import paths to `storybook/internal/instrumenter`
- Removed `@storybook/instrumenter` dependencies from multiple packages as functionality moves to core
- Build size increases noted: buildSbAddonsSize (+81kB, +4.1%) and overall buildSize (+119kB, +1.6%)
- Significant createTime performance regression (+70.3%, +18.8s) warrants investigation
- Added new exports and types in core/package.json for instrumenter module under 'internal/instrumenter'



<!-- /greptile_comment -->